Expose Trials in env-file. Improving code. (#12)

+++ scripts/training/stop.sh
@@ -2,5 +2,9 @@
 
 docker-compose down
 
-docker stop $(docker ps | awk ' /sagemaker/ { print $1 }')
-docker rm $(docker ps -a | awk ' /sagemaker/ { print $1 }')
+SAGEMAKER=$(docker ps | awk ' /sagemaker/ { print $1 }')
+if [[ -n $SAGEMAKER ]];
+then
+    docker stop $(docker ps | awk ' /sagemaker/ { print $1 }')
+    docker rm $(docker ps -a | awk ' /sagemaker/ { print $1 }')
+fi
